Modularize pipeline action functions. Add attr() action function
